Sid2

See merge request courses/2022-winter/csci-5308/group2/parkingpool!11

+++ com.parkingpool.backend/src/main/java/Modules/Review/model/Review.java
@@ -5,8 +5,8 @@ public class Review {
     public int userID;
     public int parkingID;
     public String reviews;
-    public int ratings;
-    public Review(int review_id, int userID, int parkingID, String reviews, int ratings)
+    public double ratings;
+    public Review(int review_id, int userID, int parkingID, String reviews, double ratings)
     {
         this.review_id = review_id;
         this.userID = userID;
